Refactored code based on review feedback and reorganized test structure
