+++ src/lib.rs
@@ -11,12 +11,16 @@
 // ---
 #![cfg_attr(docsrs, feature(doc_cfg))]
 
+/// zuc algorithms module
 mod zuc;
-pub use self::zuc::ZUC;
-mod zuc128;
-pub use self::zuc128::ZUC128;
-mod zuc256;
-pub use self::zuc256::{MacLength, ZUC256};
 
+/// zuc 128 bit
+mod zuc128;
+pub use self::zuc128::Zuc128;
+
+/// zuc 256bit
+mod zuc256;
+pub use self::zuc256::{MacLength, Zuc256};
+
+/// utils for bit calculate
 mod utils;
-mod zuc_data;
