+++ src/zuc128.rs
@@ -1,37 +1,57 @@
 //! ZUC-128 Algorithms
 
-use super::zuc::ZUC;
-use super::zuc_data::D_128;
+use crate::zuc::Zuc;
+
+/// d constants 128
+pub static D_128: [u16; 16] = [
+    0b_0100_0100_1101_0111,
+    0b_0010_0110_1011_1100,
+    0b_0110_0010_0110_1011,
+    0b_0001_0011_0101_1110,
+    0b_0101_0111_1000_1001,
+    0b_0011_0101_1110_0010,
+    0b_0111_0001_0011_0101,
+    0b_0000_1001_1010_1111,
+    0b_0100_1101_0111_1000,
+    0b_0010_1111_0001_0011,
+    0b_0110_1011_1100_0100,
+    0b_0001_1010_1111_0001,
+    0b_0101_1110_0010_0110,
+    0b_0011_1100_0100_1101,
+    0b_0111_1000_1001_1010,
+    0b_0100_0111_1010_1100,
+];
 
 /// ZUC128 keystream generator
 /// ([GB/T 33133.1-2016](https://openstd.samr.gov.cn/bzgk/gb/newGbInfo?hcno=8C41A3AEECCA52B5C0011C8010CF0715))
 #[derive(Debug, Clone)]
-pub struct ZUC128 {
-    zuc: ZUC,
+pub struct Zuc128 {
+    /// zuc core
+    core: Zuc,
 }
 
-impl ZUC128 {
+impl Zuc128 {
     /// Creates a ZUC128 keystream generator
     #[must_use]
     pub fn new(key: &[u8; 16], iv: &[u8; 16]) -> Self {
-        let mut zuc = ZUC::zeroed();
-        let mut s: [u32; 16] = [0; 16];
+        let mut zuc = Zuc::zeroed();
         for i in 0..16 {
             let k_i = u32::from(key[i]);
             let d_i = u32::from(D_128[i]);
             let iv_i = u32::from(iv[i]);
-            s[i] = (k_i << 23) | (d_i << 8) | iv_i;
+            zuc.s[i] = (k_i << 23) | (d_i << 8) | iv_i;
         }
-        zuc.init(s);
-        Self { zuc }
+        zuc.init();
+        Self { core: zuc }
     }
+
     ///  Generates the next 32-bit word in ZUC128 keystream
     pub fn generate(&mut self) -> u32 {
-        self.zuc.generate()
+        self.core.generate()
     }
 }
 
-impl Iterator for ZUC128 {
+impl Iterator for Zuc128 {
     type Item = u32;
 
     #[inline]
@@ -39,10 +59,12 @@ impl Iterator for ZUC128 {
         Some(self.generate())
     }
 }
+
 #[cfg(test)]
 mod tests {
-    use crate::ZUC128;
+    use crate::Zuc128;
 
+    // examples from http://c.gb688.cn/bzgk/gb/showGb?type=online&hcno=8C41A3AEECCA52B5C0011C8010CF0715
     struct Example {
         k: [u8; 16],
         iv: [u8; 16],
@@ -167,30 +189,30 @@ mod tests {
     };
 
     #[test]
-    fn examples() {
+    fn unit_test_zuc_128() {
         for Example { k, iv, expected } in [&EXAMPLE1, &EXAMPLE2, &EXAMPLE3] {
-            let mut zuc = ZUC128::new(k, iv);
+            let mut zuc = Zuc128::new(k, iv);
 
-            assert_eq!(zuc.zuc.x, expected[0][..4]);
-            assert_eq!(zuc.zuc.r1, expected[0][4]);
-            assert_eq!(zuc.zuc.r2, expected[0][5]);
-            assert_eq!(zuc.zuc.s[15], expected[0][7]);
+            assert_eq!(zuc.core.x, expected[0][..4]);
+            assert_eq!(zuc.core.r1, expected[0][4]);
+            assert_eq!(zuc.core.r2, expected[0][5]);
+            assert_eq!(zuc.core.s[15], expected[0][7]);
 
             let z1 = zuc.generate();
 
-            assert_eq!(zuc.zuc.x, expected[1][..4]);
-            assert_eq!(zuc.zuc.r1, expected[1][4]);
-            assert_eq!(zuc.zuc.r2, expected[1][5]);
+            assert_eq!(zuc.core.x, expected[1][..4]);
+            assert_eq!(zuc.core.r1, expected[1][4]);
+            assert_eq!(zuc.core.r2, expected[1][5]);
             assert_eq!(z1, expected[1][6]);
-            assert_eq!(zuc.zuc.s[15], expected[1][7]);
+            assert_eq!(zuc.core.s[15], expected[1][7]);
 
             let z2 = zuc.generate();
 
-            assert_eq!(zuc.zuc.x, expected[2][..4]);
-            assert_eq!(zuc.zuc.r1, expected[2][4]);
-            assert_eq!(zuc.zuc.r2, expected[2][5]);
+            assert_eq!(zuc.core.x, expected[2][..4]);
+            assert_eq!(zuc.core.r1, expected[2][4]);
+            assert_eq!(zuc.core.r2, expected[2][5]);
             assert_eq!(z2, expected[2][6]);
-            assert_eq!(zuc.zuc.s[15], expected[2][7]);
+            assert_eq!(zuc.core.s[15], expected[2][7]);
         }
     }
 }
